A detached section.
Proposal: It may be more sensible to allow computers within the gameshow to communicate on a peer-to-peer basis (no server, just independent computers connecting to each other with one being the 'referee,' or the 'regulator' as written in the program) using UDP (not too reliable, but it works). Hopefully this is good enough.